requirements.txt
langchain==0.3.26
flask==3.1.1
sentence-transformers==4.1.0
pypdf==5.6.1
jsonloader==0.9.1
jq==1.10.0
python-dotenv==1.1.0
langchain-pinecone==0.2.8
langchain-google-genai==2.1.8
langchain-community==0.3.26
-e .